Despesa recorrente finalizado.

+++ app/Repositories/Rules/Arquivo.php
@@ -12,7 +12,6 @@ class Arquivo
     public static function create(DespesaRecorrente|User $despesa, array $data, string $tipo):void
     {
         $params = ["id_despesa_recorrente" => $despesa->id, "tipo_documento" => $tipo];
-        $deleteFile = $despesa->arquivos->where("id_despesa_recorrente", $despesa->id)->where("tipo_documento", $tipo)->first();
 
         $file = $despesa->arquivos()->updateOrCreate($params, [
             'tipo' => $data[$tipo]->getMimeType(),
@@ -22,11 +21,22 @@ class Arquivo
             'tipo_documento' => $tipo,
         ]);
 
+        self::deleteFile($despesa, $tipo);
+        self::storeFileAndSave($data, $tipo, $file, $despesa);
+    }
 
-        Storage::disk('local')->delete('arquivos/'.$deleteFile->id.'.'.$deleteFile->extensao);
+    private static function deleteFile($despesa, string $tipo):void
+    {
+        $deleteFile = $despesa->arquivos->where("id_despesa_recorrente", $despesa->id)->where("tipo_documento", $tipo)->first();
+        if($deleteFile) {
+            Storage::disk('local')->delete('arquivos/' . $deleteFile->id . '.' . $deleteFile->extensao);
+        }
+    }
+
+    private static function storeFileAndSave(array $data, string $tipo, $file, $despesa):void
+    {
         $arquivo = $data[$tipo]->storeAs('arquivos', $file->id.'.'.$data[$tipo]->getClientOriginalExtension());
-
-        $despesa->$tipo = $file->id.'.'.$data[$tipo]->getClientOriginalExtension();
+        $despesa->$tipo = $file->id;
         $despesa->save();
     }
 }
